nonebot-plugin-puzzle/pyproject.toml
[project]
name = "nonebot-plugin-puzzle"
version = "1.0.4"
description = "在QQ群玩数字华容道，基于nonebot框架的数字华容道插件，plugin of puzzle base on nonebot frame"
authors = [
    {name = "initialencounter", email = "2911583893@qq.com"},
]
license = "AGPL-3.0"
dependencies = ["nonebot2>=2.0.0b2", "nonebot-adapter-onebot>=2.0.0b1", "Pillow>=9.0.0", "puzzle-rs>=0.2.0" ]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/initialencounter/puzzle"
Repository = "https://github.com/initialencounter/puzzle"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
